excercise 4.21, fix 4.20

+++ bloglist-app/controllers/blogs.js
@@ -30,12 +30,12 @@ blogsRouter.post('/', async (request, response, next) => {
 
   try {
     const decodedToken = jwt.verify(request.token, process.env.SECRET)
-    if (!token || !decodedToken.id) {
+
+    if (!request.token || !decodedToken.id) {
       return response.status(401).json({ error: 'token missing or invalid' })
     }
 
     const user = await User.findById(decodedToken.id)
-
     const blog = new Blog({
       ...body,
       likes: body.likes === undefined ? 0 : body.likes,
@@ -72,8 +72,21 @@ blogsRouter.put('/:id', async (request, response, next) => {
 
 blogsRouter.delete('/:id', async (request, response, next) => {
   try {
-    await Blog.findByIdAndRemove(request.params.id)
-    response.status(204).end()
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!request.token || !decodedToken.id) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
+
+    const user = await User.findById(decodedToken.id)
+    const blog = await Blog.findById(request.params.id)
+  
+    if (blog.user.toString() === user.id.toString()) {
+      await Blog.findByIdAndRemove(request.params.id)
+      response.status(204).end()
+    }
+
+    return response.status(401).json({ error: 'invalid token' })
+
   } catch (exception) {
     next(exception)
   }
